requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
GDAL>=3.1
geopandas==0.13.2 
gunicorn==20.1.0
pytorch==2.0.1  
torchvision==0.15.2  
rasterio==1.3.6 
requests==2.31.0
Pillow==9.5.0
numpy==1.24.1
albumentations==1.3.1
shapely==2.0.1